authentication/app/main.py
from typing import Annotated
from shared.models.token import Token
from fastapi.security import OAuth2PasswordRequestForm, OAuth2PasswordBearer
from datetime import datetime, timedelta, timezone
from fastapi import FastAPI, Depends, status, HTTPException
from app import config
from jose import JWTError, jwt

oauth2_authentication = OAuth2PasswordBearer(tokenUrl="token")
app = FastAPI()
@app.post("/generate_token", response_model=Token)
def login(form_data: Annotated[OAuth2PasswordRequestForm, Depends()]):
    try:
        access_token_expires = timedelta(minutes=config.ACCESS_TOKEN_EXPIRE_MINUTES)
        access_token = create_access_token(
            data={"sub": form_data.username}, expires_delta=access_token_expires
        )
        
        return Token(
            access_token=access_token,
            token_type="bearer"
        )
    except Exception as e:
        raise HTTPException(
            status_code=status.HTTP_200_OK,
            detail=str(e),
            headers={"WWW-Authenticate": "Bearer"}
        )
        
def create_access_token(data: dict, expires_delta: timedelta | None = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now(timezone.utc) + expires_delta
    else:
        expire = datetime.now(timezone.utc) + timedelta(minutes=15)
        
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, config.SECRET_KEY, algorithm=config.ALGORITHM)
    return encoded_jwt

def decode_access_token(token: str):
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, config.SECRET_KEY, algorithms=[config.ALGORITHM])
        username = str(payload.get("sub"))
        return username
        # user = get_user(fake_users_db, username)
        # if user is None:
        #     raise credentials_exception
        # token_data = TokenData(username=username)
        # return token_data
    except JWTError:
        raise credentials_exception