authentication/pyproject.toml
[tool.poetry]
name = "authentication"
version = "0.1.0"
description = ""
authors = ["bilalmk <bilalmk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.7.4"
python-jose = "^3.3.0"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
